Moved custom deliveries to separate file naively and import naively
