+++ packages/react-components/lib/tasks/descriptions/utils.ts
@@ -1,0 +1,73 @@
+import { TaskRequest } from 'api-client';
+import { GoToPlaceActivity, LotPickupActivity } from './delivery_custom';
+
+export function isNonEmptyString(value: string): boolean {
+  return value.length > 0;
+}
+
+export function isPositiveNumber(value: number): boolean {
+  return value > 0;
+}
+
+export function getShortDescription(taskRequest: TaskRequest): string | undefined {
+  if (taskRequest.category === 'patrol') {
+    const formattedPlaces = taskRequest.description.places.map((place: string) => `[${place}]`);
+    return `[Patrol] [${taskRequest.description.rounds}] round/s, along ${formattedPlaces.join(
+      ', ',
+    )}`;
+  }
+
+  // This section is only valid for custom delivery types
+  // FIXME: This block looks like it makes assumptions about the structure of
+  // the task description in order to parse it, but it is following the
+  // statically defined description (object) at the top of this file. The
+  // descriptions should be replaced by a schema in general, however the better
+  // approach now should be to make each task description testable and in charge
+  // of their own short descriptions.
+  try {
+    const goToPickup: GoToPlaceActivity =
+      taskRequest.description.phases[0].activity.description.activities[0];
+    const pickup: LotPickupActivity =
+      taskRequest.description.phases[0].activity.description.activities[1];
+    const cartId = pickup.description.description.cart_id;
+    const goToDropoff: GoToPlaceActivity =
+      taskRequest.description.phases[1].activity.description.activities[0];
+
+    switch (taskRequest.description.category) {
+      case 'delivery_pickup': {
+        return `[Delivery - 1:1] payload [${cartId}] from [${goToPickup.description}] to [${goToDropoff.description}]`;
+      }
+      case 'delivery_sequential_lot_pickup': {
+        return `[Delivery - Sequential lot pick up] payload [${cartId}] from [${goToPickup.description}] to [${goToDropoff.description}]`;
+      }
+      case 'delivery_area_pickup': {
+        return `[Delivery - Area pick up] payload [${cartId}] from [${goToPickup.description}] to [${goToDropoff.description}]`;
+      }
+      default:
+        return `[Unknown] type "${taskRequest.description.category}"`;
+    }
+  } catch (e) {
+    if (e instanceof TypeError) {
+      console.error(`Failed to parse custom delivery: ${e.message}`);
+    } else {
+      console.error(
+        `Failed to generate short description from task of category: ${taskRequest.category}: ${
+          (e as Error).message
+        }`,
+      );
+    }
+
+    try {
+      const descriptionString = JSON.stringify(taskRequest.description);
+      console.error(descriptionString);
+      return descriptionString;
+    } catch (e) {
+      console.error(
+        `Failed to parse description of task of category: ${taskRequest.category}: ${
+          (e as Error).message
+        }`,
+      );
+      return undefined;
+    }
+  }
+}
