+++ packages/react-components/lib/tasks/index.ts
@@ -1,4 +1,5 @@
 export * from './create-task';
+export * from './descriptions';
 export * from './task-info';
 export * from './task-logs';
 export * from './task-table';
